build wavelet manager
supporting number of amplitudes

added plot_summary

started some experiments

good place regarding the wavelet decomposition

+++ Signal/Signal.py
@@ -15,6 +15,6 @@ class Signal:
                f'\tphase={self.phase}\n'
 
     def evaluate(self, duration, samples_per_second):
-        sampling_times = np.array([i / samples_per_second for i in range(duration * samples_per_second)])
+        sampling_times = np.array([i / samples_per_second for i in range(int(duration * samples_per_second))])
         evaluated_signal = self.amplitude * np.sin(2 * np.pi * self.frequency * sampling_times)
         return evaluated_signal
